Refactor lookup calls to unify data retrieval methods across the codebase
