+++ libs/data/__init__.py
@@ -5,5 +5,6 @@
 - `libs.data.aggregate`: データ集計
 - `libs.data.initialization`: データベース初期化
 - `libs.data.loader`: SQLクエリ実行
+- `libs.data.lookup`: 情報取得
 - `libs.data.modify`: DBデータ操作
 """
