+++ libs/registry/team.py
@@ -36,7 +36,7 @@ def create(argument: list) -> str:
                 )
                 resultdb.commit()
                 resultdb.close()
-                g.cfg.team.info = lookup.db.get_team_info()
+                g.cfg.team.info = lookup.get_team_info()
                 msg = f"チーム「{team_name}」を登録しました。"
                 logging.info("add new team: %s", team_name)
 
@@ -70,7 +70,7 @@ def delete(argument: list) -> str:
             )
             resultdb.commit()
             resultdb.close()
-            g.cfg.team.info = lookup.db.get_team_info()
+            g.cfg.team.info = lookup.get_team_info()
             msg += f"\nチーム「{team_name}」を削除しました。"
             logging.info("team delete: %s", team_name)
 
@@ -128,7 +128,7 @@ def append(argument: list) -> str:
             )
             resultdb.commit()
             resultdb.close()
-            g.cfg.team.info = lookup.db.get_team_info()
+            g.cfg.team.info = lookup.get_team_info()
             msg = f"チーム「{team_name}」に「{player_name}」を所属させました。"
             logging.info("team participation: %s -> %s", team_name, player_name)
 
@@ -180,7 +180,7 @@ def remove(argument: list) -> str:
             )
             resultdb.commit()
             resultdb.close()
-            g.cfg.team.info = lookup.db.get_team_info()
+            g.cfg.team.info = lookup.get_team_info()
             msg = f"チーム「{team_name}」から「{player_name}」を離脱させました。"
             logging.info("team breakaway: %s -> %s", team_name, player_name)
 
@@ -204,7 +204,7 @@ def clear() -> str:
     resultdb.close()
 
     initialization.initialization_resultdb(g.cfg.setting.database_file)
-    g.cfg.member.info = lookup.db.get_member_info()
-    g.cfg.team.info = lookup.db.get_team_info()
+    g.cfg.member.info = lookup.get_member_info()
+    g.cfg.team.info = lookup.get_team_info()
 
     return msg
