+++ tests/name_check/test_name.py
@@ -23,7 +23,7 @@ def test_name_permit(input_args, expected_flags, monkeypatch):
     """メンバー登録テスト(OK)"""
     monkeypatch.setattr(sys, "argv", TEST_ARGS)
     configuration.setup(init_db=False)
-    lookup.db.read_memberslist()
+    lookup.read_memberslist()
 
     flg, reason = validator.check_namepattern(input_args, "member")
     print(" -->", flg, reason)
@@ -39,7 +39,7 @@ def test_name_refusal(input_args, expected_flags, monkeypatch):
     """メンバー登録テスト(NG)"""
     monkeypatch.setattr(sys, "argv", TEST_ARGS)
     configuration.setup(init_db=False)
-    lookup.db.read_memberslist()
+    lookup.read_memberslist()
 
     flg, _ = validator.check_namepattern(input_args, "member")
     assert flg == expected_flags
